services/worker: run executor with client's cache/workspace directory paths
aids e.g. debugging by having identical filesystem paths across the
distributed build. bubblewrap must be present in the worker image in
order to create the required Linux namespace. if bubblewrap is missing
or the worker is running on Windows, the worker's default and likely
mismatching paths will be used.

Change-Id: Ia50dd9f3e04d36f8fc23168a8d032413115e8625
Signed-off-by: Robert Andersson <robert.m.andersson@atlascopco.com>