tasks: introduce new task and resource attribute 'local'
In a network build, a local task or resource is always run in the
client and never in a worker. This is particularly useful for
blocking resource acquisitions that would otherwise occupy
workers and prevent execution of ready tasks.

The change also introduces resource artifacts that are published
and uploaded to the global cache, also if resource acquisition
fails. As with tasks, multiple artifacts are supported through
the attributes.artifact decorator.

Signed-off-by: Robert Andersson <robert.m.andersson@atlascopco.com>